pyproject.toml
[project]
name = "med-rag-bot"
version = "0.1.0"
description = "Medical RAG Chatbot based on a medical textbook"
authors = [
    {name = "vansh-561",email = "vanshagrawal561@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pypdf (==5.5.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tqdm (==4.67.1)",
    "tiktoken (==0.9.0)",
    "langchain (>=0.1.4)",
    "langchain-community (>=0.3.23)",
    "google-generativeai (>=0.4.0)",
    "pinecone (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "med_rag_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
